+++ test/test.carbone.js
@@ -1003,6 +1003,203 @@ describe('Carbone', function () {
         });
       });
     });
+
+    describe('security test', function () {
+      it('cannot inject JS using XML', function (done) {
+        var data = {
+          param : 3,
+        };
+        carbone.renderXML("<xml>{d.param}\\'sdsdsd.ss;'{d.param} </xml>", data, function (err, result) {
+          helper.assert(err+'', 'null');
+          helper.assert(result, "<xml>3\\\\'sdsdsd.ss;3\\' </xml>");
+          done();
+        });
+      });
+      it('should accept non-alphanumeric characters in variable names', function (done) {
+        var data = {
+          o           : { id : 2 },
+          'r🚀cket'   : { id : 3 },
+          'qu\\\'ote' : { id : 4 },
+          报道          : { id : 5 }
+        };
+        carbone.renderXML('<xml>{d.o.id} ha {d.r🚀cket.id} he {d.qu\\\'ote.id} ch {d.报道.id}</xml>', data, function (err, result) {
+          helper.assert(err+'', 'null');
+          helper.assert(result, '<xml>2 ha 3 he 4 ch 5</xml>');
+          done();
+        });
+      });
+      it('should accept non-alphanumeric characters in arrays', function (done) {
+        var data = [{
+          o           : { id : 2 },
+          'r🚀cket'   : { id : 3 },
+          'qu\\\'ote' : { id : 4 },
+          报道          : { id : 5 }
+        }];
+        carbone.renderXML('<xml><tr>{d[i].o.id} ha {d[i].r🚀cket.id} he {d[i].qu\\\'ote.id} ch {d[i].报道.id}</tr><tr>{d[i+1].o.id}</tr></xml>', data, function (err, result) {
+          helper.assert(err+'', 'null');
+          helper.assert(result, '<xml><tr>2 ha 3 he 4 ch 5</tr></xml>');
+          done();
+        });
+      });
+      it('should accept non-alphanumeric characters in arrays conditions', function (done) {
+        var data = [{
+          o           : { id : 2, 'i💎d' : 200, 'i\\\'d' : 2000 },
+          'r🚀cket'   : { id : 3, 'i💎d' : 300, 'i\\\'d' : 3000 },
+          'qu\\\'ote' : { id : 4, 'i💎d' : 400, 'i\\\'d' : 4000 },
+          报道          : { id : 5, 'i💎d' : 500, 'i\\\'d' : 5000 },
+          'qu"ote'    : { id : 6, 'i💎d' : 600, 'i"d' : 600 }
+        },
+        {
+          o           : { id : 12, 'i💎d' : 1200, 'i\\\'d' : 1200 },
+          'r🚀cket'   : { id : 13, 'i💎d' : 1300, 'i\\\'d' : 1300 },
+          'qu\\\'ote' : { id : 14, 'i💎d' : 1400, 'i\\\'d' : 1400 },
+          报道          : { id : 15, 'i💎d' : 1500, 'i\\\'d' : 1500 },
+          'qu"ote'    : { id : 16, 'i💎d' : 1600, 'i"d' : 1600 }
+        }];
+        carbone.renderXML('<xml><tr>{d[i, r🚀cket.id = 3 ].o.id}</tr><tr>{d[i+1, r🚀cket.id = 3].o.id}</tr></xml>', data, function (err, result) {
+          helper.assert(err+'', 'null');
+          helper.assert(result, '<xml><tr>2</tr></xml>');
+          carbone.renderXML('<xml><tr>{d[i, r🚀cket.i💎d = 300 ].o.id}</tr><tr>{d[i+1, r🚀cket.i💎d = 300].o.id}</tr></xml>', data, function (err, result) {
+            helper.assert(err+'', 'null');
+            helper.assert(result, '<xml><tr>2</tr></xml>');
+            carbone.renderXML('<xml><tr>{d[i, r🚀cket.i\\\'d = 3000 ].o.id}</tr><tr>{d[i+1, r🚀cket.i\\\'d = 3000].o.id}</tr></xml>', data, function (err, result) {
+              helper.assert(err+'', 'null');
+              helper.assert(result, '<xml><tr>2</tr></xml>');
+              carbone.renderXML('<xml><tr>{d[i, qu\\\'ote.id = 4 ].o.id}</tr><tr>{d[i+1, qu\\\'ote.id = 4].o.id}</tr></xml>', data, function (err, result) {
+                helper.assert(err+'', 'null');
+                helper.assert(result, '<xml><tr>2</tr></xml>');
+                carbone.renderXML('<xml><tr>{d[i, 报道.id = 5 ].o.id}</tr><tr>{d[i+1, 报道.id = 5].o.id}</tr></xml>', data, function (err, result) {
+                  helper.assert(err+'', 'null');
+                  helper.assert(result, '<xml><tr>2</tr></xml>');
+                  carbone.renderXML('<xml><tr>{d[i, qu"ote.i"d = 600 ].qu"ote.i"d}</tr><tr>{d[i+1, qu"ote.i"d = 600].qu"ote.i"d}</tr></xml>', data, function (err, result) {
+                    helper.assert(err+'', 'null');
+                    helper.assert(result, '<xml><tr>600</tr></xml>');
+                    done();
+                  });
+                });
+              });
+            });
+          });
+        });
+      });
+      it('should accept non-alphanumeric characters in conditions values', function (done) {
+        var data = [
+          { o : { id : 'i💎d'   }, val : 'i💎d'  , res : 1 },
+          { o : { id : 'i\\\'d' }, val : 'i\\\'d', res : 2 },
+          { o : { id : '报道'    }, val : '报道'   , res : 3 },
+          { o : { id : 'tes\\"' }, val : 'tes\\"', res : 4 },
+          { o : { id : 'at"'    }, val : 'at"'   , res : 5 }
+        ];
+        carbone.renderXML('<xml><tr>{d[i, o.id = i💎d, val = i💎d ].res}</tr><tr>{d[i+1, o.id = i💎d, val = i💎d].res}</tr></xml>', data, function (err, result) {
+          helper.assert(err+'', 'null');
+          helper.assert(result, '<xml><tr>1</tr></xml>');
+          carbone.renderXML('<xml><tr>{d[i, o.id = i\\\'d, val = i\\\'d ].res}</tr><tr>{d[i+1, o.id = i\\\'d, val = i\\\'d ].res}</tr></xml>', data, function (err, result) {
+            helper.assert(err+'', 'null');
+            helper.assert(result, '<xml><tr>2</tr></xml>');
+            carbone.renderXML('<xml><tr>{d[i, o.id = 报道, val = 报道 ].res}</tr><tr>{d[i+1, o.id = 报道, val = 报道 ].res}</tr></xml>', data, function (err, result) {
+              helper.assert(err+'', 'null');
+              helper.assert(result, '<xml><tr>3</tr></xml>');
+              carbone.renderXML('<xml><tr>{d[i, o.id = "tes\\"", val = "tes\\""].res}</tr><tr>{d[i+1, o.id = "tes\\"", val = "tes\\""].res}</tr></xml>', data, function (err, result) {
+                helper.assert(err+'', 'null');
+                helper.assert(result, '<xml><tr>4</tr></xml>');
+                carbone.renderXML('<xml><tr>{d[i, o.id = "at"",val=  "at""].res}</tr><tr>{d[i+1, o.id = "at"",val= "at""].res}</tr></xml>', data, function (err, result) {
+                  helper.assert(err+'', 'null');
+                  helper.assert(result, '<xml><tr>5</tr></xml>');
+                  done();
+                });
+              });
+            });
+          });
+        });
+      });
+      it('should work even with antislash and quotes', function (done) {
+        var data = {
+          "p\\\\'aram" : 3,
+        };
+        carbone.renderXML("<xml>{d.p\\\\'aram}</xml>", data, function (err, result) {
+          helper.assert(err+'', 'null');
+          helper.assert(result, '<xml>3</xml>');
+          done();
+        });
+      });
+      it('should work even if there are quotes in XML and attributes', function (done) {
+        var data = {
+          "p\\\\'aram" : 3,
+        };
+        carbone.renderXML("<xml>{d.p\\\\'aram}' </xml>", data, function (err, result) {
+          helper.assert(err+'', 'null');
+          helper.assert(result, '<xml>3\\\' </xml>');
+          done();
+        });
+      });
+      it.skip('should work even if there are quotes in XML', function (done) {
+        var data = {
+          param : 3,
+        };
+        carbone.renderXML("<a>';]<xml>{d.param}' </xml>", data, function (err, result) {
+          helper.assert(err+'', 'null');
+          helper.assert(result, "<a>\\';]<xml>3\\' </xml>");
+          done();
+        });
+      });
+      it.skip('should work even if there are quotes in XML', function (done) {
+        var data = {
+          param : 3,
+        };
+        carbone.renderXML("<a>\\';]<xml>{d.param}' </xml>", data, function (err, result) {
+          helper.assert(err+'', 'null');
+          helper.assert(result, "<a>\\\\';]<xml>3\\' </xml>");
+          done();
+        });
+      });
+      it('should not crash if someone tries to inject JS code in variable', function (done) {
+        var _xml = "{#myVar= '];console%2log(sd)'//  }<xml> <t_row> {$myVar} </t_row> </xml>";
+        var _data = [
+          {brand : 'Lumeneo'     , id : 1},
+          {brand : 'Tesla motors', id : 2},
+          {brand : 'Toyota'      , id : 3}
+        ];
+        carbone.renderXML(_xml, _data, function (err, _xmlBuilt) {
+          helper.assert(err+'', 'null');
+          helper.assert(_xmlBuilt, '<xml> <t_row>  </t_row> </xml>');
+          done();
+        });
+      });
+      it('should not crash if weird quotes are injected in formatter parameters', function (done) {
+        var data = {};
+        carbone.renderXML('<xml>{d:ifEmpty(\'\\\\\'yeah\')}</xml>', data, {complement : {}}, function (err, result) {
+          helper.assert(err+'', 'null');
+          helper.assert(result, '<xml>\\\\\'yeah</xml>');
+          done();
+        });
+      });
+      it('should not crash if the iterator contains string and is negative', function (done) {
+        var _xml = '<xml> <t_row> {d[i=-1a].brand} </t_row><t_row> {d[i=-2a].brand} </t_row></xml>';
+        var _data = [
+          {brand : 'Lumeneo'     , id : 1},
+          {brand : 'Tesla motors', id : 2},
+          {brand : 'Toyota'      , id : 3}
+        ];
+        carbone.renderXML(_xml, _data, function (err, _xmlBuilt) {
+          helper.assert(err+'', 'null');
+          helper.assert(_xmlBuilt, '<xml> <t_row> Toyota </t_row><t_row> Tesla motors </t_row></xml>');
+          done();
+        });
+      });
+      it('should not crash if the iterator contains string', function (done) {
+        var _xml = '<xml> <t_row> {d[i=1a].brand} </t_row><t_row> {d[i=2a].brand} </t_row></xml>';
+        var _data = [
+          {brand : 'Lumeneo'     , id : 1},
+          {brand : 'Tesla motors', id : 2},
+          {brand : 'Toyota'      , id : 3}
+        ];
+        carbone.renderXML(_xml, _data, function (err, _xmlBuilt) {
+          helper.assert(err+'', 'null');
+          helper.assert(_xmlBuilt, '<xml> <t_row> Tesla motors </t_row><t_row> Toyota </t_row></xml>');
+          done();
+        });
+      });
+    });
     describe('Conditional block and conditions formatters showBegin/showEnd', function () {
       it('should accept to use other formatters with conditional blocks', function (done) {
         var _xml = '<xml> {d.val:ifEQ(3):hideBegin} <a></a> {d.val:hideEnd} {d.val:ifEQ(2):hideBegin} <b></b> {d.val:hideEnd} </xml>';