+++ CHANGELOG.md
@@ -126,6 +126,8 @@
              `{d.cars[i, type="Tesla car"].name}`
 
   - (Fix LibreOffice detection on Windows)
+  - Accepts non-alphanumeric characters in variables names, values, ... For example, `{d.i💎d}` is allowed
+  - Improve security in the builder and reduce memory consumption
 
 
 ### v1.2.1